requirements.txt
fastapi==0.85.0
uvicorn==0.18.3
motor==3.3.1
pymongo==4.5.0
python-dotenv==0.21.1
requests==2.28.1
beautifulsoup4==4.11.1
SQLAlchemy==2.0
pydantic>=1.9.0
pytest==8.3.5
anyio==3.6.1
cloudscraper==1.2.71
pytest-asyncio
playwright==1.52.0